aesthetic_eval_score_framework/requirements-common.txt
numpy>=1.26
Pillow>=10.0
PyYAML>=6.0
tqdm>=4.66
scipy>=1.11
packaging>=23.0
